fix(python/sdk): change sentence a paragraph confidence type to `float` (#3384)
GitOrigin-RevId: eb4cdb766c247dba4d24325d5a01057148186188

+++ assemblyai/types.py
@@ -1385,7 +1385,7 @@ class Sentence(Word):
     words: List[Word]
     start: int
     end: int
-    confidence: int
+    confidence: float
     speaker: Optional[str]
 
 
@@ -1399,7 +1399,7 @@ class Paragraph(Word):
     words: List[Word]
     start: int
     end: int
-    confidence: int
+    confidence: float
     text: str
